sales_command_center/requirements.txt
# Sales Command Center - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI & LLM
openai==1.3.5
anthropic==0.7.0
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Cache
redis==5.0.1
hiredis==2.2.3

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# CRM/ERP Integrations
simple-salesforce==1.12.5

# Task Queue (optional)
celery==5.3.4
kombu==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk==1.38.0

# AWS SDK (for S3, SES, etc.)
boto3==1.34.0
botocore==1.34.0

# Email
sendgrid==6.11.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx-mock==0.14.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2

# Utilities
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# CORS
fastapi-cors==0.0.6

# WebSockets
websockets==12.0

# Rate Limiting
slowapi==0.1.9

# JSON Web Tokens
pyjwt==2.8.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Charts & Visualization (backend)
matplotlib==3.8.2
seaborn==0.13.0

# PDF Generation (for reports)
reportlab==4.0.7
weasyprint==60.1

# Excel Export
openpyxl==3.1.2
xlsxwriter==3.1.9
